more heavily utilised the mod.rs files for storing code in less obscure places
