+++ server/src/core/mod.rs
@@ -5,5 +5,5 @@ mod search;
 
 pub fn configure(cfg: &mut web::ServiceConfig) {
     cfg.service(get::get_handler);
-    cfg.service(search::search_handler::search_handler);
+    cfg.service(search::search_handler);
 }
